04 - Chat App/backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-environ==0.9.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
